Второе ревью 14-го спринта
Добавить проверку пользователя в UserServiceImpl.deleteUser

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -50,6 +50,9 @@ class UserServiceImpl implements UserService {
 
 	@Override
 	public void deleteUser(long userId) {
+		if (userRepository.checkUserIsNotPresent(userId)) {
+			throw new NotFoundException("Пользователь с id=" + userId + " не найден.");
+		}
 		userRepository.remove(userId);
 	}
